processes
I created the initial code for processes, but it is incomplete.  It needs more error checking, and I don't know how to get os161 to recognize them.  This was why I couldn't get hello.c to work in assignment 0.  Note that I also made minor changes to thread.h and thread.c to accomodate processes.  Be sure to take a look at that.  I need to take a nap, since I can't think straight right now. Best of luck.